Fix connect to server

+++ app/src/main/java/com/example/vinid_icecreams/model/Store.kt
@@ -5,8 +5,8 @@ data class Store(
     var name:String,
     var address: String,
     var image :String,
-    var latitude: Int,
-    var longitude: Int,
+    var latitude: Double,
+    var longitude: Double,
     var createdAt: Long,
     var mListIceCream: ArrayList<IceCream>
 )
